troubleshoot: dx network request stall during vite websocket setup
