+++ blogcms-react/src/pages/AuthPage.tsx
@@ -15,55 +15,37 @@ import { useState } from 'react';
 import LoginForm from '../components/LoginForm';
 import RegisterForm from '../components/RegisterForm';
 import { sbClient } from '../services/sb';
+import { theme } from '../styles/theme';
 /**
  * Interfaces and types
  */
 
   
-  const useStyles = createStyles((theme) => ({
-    wrapper: {
-      minHeight: 900,
-      backgroundSize: 'cover',
-      backgroundImage:
-        'url(https://images.unsplash.com/photo-1484242857719-4b9144542727?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=1280&q=80)',
-    },
   
-    form: {
-      borderRight: `1px solid ${
-        theme.colorScheme === 'dark' ? theme.colors.dark[7] : theme.colors.gray[3]
-      }`,
-      minHeight: 900,
-      maxWidth: 450,
-      paddingTop: 80,
-  
-      [`@media (max-width: ${theme.breakpoints.sm}px)`]: {
-        maxWidth: '100%',
-      },
-    },
-  
-    title: {
-      color: theme.colorScheme === 'dark' ? theme.white : theme.black,
-      fontFamily: `Greycliff CF, ${theme.fontFamily}`,
-    },
-  
-    logo: {
-      color: theme.colorScheme === 'dark' ? theme.white : theme.black,
-      width: 120,
-      display: 'block',
-      marginLeft: 'auto',
-      marginRight: 'auto',
-    },
-  }));
   
   export function AuthPage() {
     const [registerPage, setRegisterPage] = useState<boolean>(false)
-    const { classes } = useStyles();
 
 
     return (
-      <div className={classes.wrapper}>
-        <Paper className={classes.form} radius={0} p={30}>
-          {registerPage === false? <LoginForm classes={classes}/> : <RegisterForm classes={classes} />}
+      <div style={{
+        minHeight: 900,
+      backgroundSize: 'cover',
+      backgroundImage:
+        'url(https://images.unsplash.com/photo-1621237023000-6a628c285938?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1740&q=80)'
+      }}>
+        <Paper style={{
+          borderRight: `1px solid ${
+      theme.colors.dark[7]}`,
+          minHeight: 900,
+          maxWidth: 450,
+          paddingTop: 80,
+      
+          [`@media (max-width: ${theme.breakpoints.sm}px)`]: {
+            maxWidth: '100%',
+          }
+        }} radius={0} p={30}>
+          {registerPage === false? <LoginForm /> : <RegisterForm />}
         </Paper>
       </div>
     );
